pyproject.toml
[tool.poetry]
name = "convert-bitcoin"
version = "0.1.0"
description = ""
authors = ["Tyler Stubenvoll <tstubenv@fastmail.fm>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
